+++ src/EventStore.Core.Tests/Helpers/MiniClusterNode.cs
@@ -63,7 +63,7 @@ namespace EventStore.Core.Tests.Helpers {
 		}
 
 		public MiniClusterNode(string pathname, int debugIndex, IPEndPoint internalTcp, IPEndPoint externalTcp,
-			IPEndPoint httpEndPoint, EndPoint[] gossipSeeds, ISubsystem[] subsystems = null,
+			IPEndPoint httpEndPoint, EndPoint[] gossipSeeds, ISubsystemFactory[] subsystems = null,
 			bool enableTrustedAuth = false, int memTableSize = 1000, bool inMemDb = true,
 			bool disableFlushToDisk = false, bool readOnlyReplica = false, int nodePriority = 0,
 			string intHostAdvertiseAs = null, IExpiryStrategy expiryStrategy = null) {
@@ -147,7 +147,7 @@ namespace EventStore.Core.Tests.Helpers {
 				Projections = new() {
 					RunProjections = ProjectionType.None
 				},
-				Subsystems = subsystems ?? Array.Empty<ISubsystem>()
+				Subsystems = subsystems ?? Array.Empty<ISubsystemFactory>()
 			};
 
 			var serverCertificate = useHttps ? ssl_connections.GetServerCertificate() : null;