+++ src/EventStore.Core/Messaging/CoreMessage.cs
@@ -21,6 +21,7 @@
 		Subscription,
 		System,
 		Tcp,
+		Telemetry,
 		Timer,
 		UserManagement
 	}