+++ src/EventStore.Core/PluginModel/ISubsystemPlugin.cs
@@ -1,9 +1,0 @@
-
-namespace EventStore.Core.PluginModel {
-	public interface ISubsystemPlugin {
-		string Name { get; }
-		string Version { get; }
-
-		ISubsystemFactory GetSubsystemFactory();
-	}
-}